+++ java/api/src/main/java/fr/ill/ics/cameo/coms/impl/SubscriberImpl.java
@@ -21,7 +21,7 @@ import fr.ill.ics.cameo.strings.Endpoint;
 
 public interface SubscriberImpl {
 	
-	void init(int appId, Endpoint endpoint, Endpoint appStatusEndpoint, String publisherIdentity) throws ConnectionTimeout;
+	void init(int appId, Endpoint endpoint, Endpoint appStatusEndpoint, String publisherIdentity, boolean checkApp) throws ConnectionTimeout;
 	boolean hasEnded();
 	boolean isCanceled();
 	byte[] receive();