+++ java/api/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Minor review of exceptions.
 * Renamed PROCESSING_ERROR into PROCESSING_FAILURE.
+* Added the checkApp feature to Requester and Subscriber classes.
 
 2.1.1
 -----